New: neustrukturierung Verzeichnisse minikaenguru (#19)
