+++ apps/minikaenguru-anwendung-ui/src/app/lehrperson/dashboard-lehrer/dashboard-lehrer.component.ts
