+++ apps/minikaenguru-anwendung-ui/src/app/privat/dashboard-privat/dashboard-privat.component.scss
