+++ apps/minikaenguru-anwendung-ui/src/app/remote.routes.ts
@@ -1,6 +1,6 @@
 import { Routes } from '@angular/router';
 import { AppComponent } from './app.component';
-import { StartComponent } from './start/start.component';
+import { StartComponent } from './start/start/start.component';
 import { mkaAuthorizationDataProvider } from './core/authorization/authorization-api/mka-authorization-data.provider';
 import { minikaenguruAnwendungConfiguration } from './config/configuration';
 import { MINIKAENGURU_ANWENDUNG_CONFIGURATION } from './config/minikaenguru-anwendung.configuration';
