+++ apps/minikaenguru-anwendung-ui/src/app/schule/dashboard-schule/dashboard-schule.component.ts
