+++ apps/minikaenguru-anwendung-ui/src/app/start/guest-info/guest-info.component.spec.ts
